First part of the translation improvements. Small code refactoring improvements.
