+++ src/Application/Features/Loggers/Commands/Clear/ClearLogsCommand.cs
@@ -4,44 +4,40 @@
 
 using CleanArchitecture.Blazor.Application.Features.Loggers.Caching;
 
-namespace CleanArchitecture.Blazor.Application.Features.Loggers.Commands.Delete;
+namespace CleanArchitecture.Blazor.Application.Features.Loggers.Commands.Clear;
 
 public class ClearLogsCommand : ICacheInvalidatorRequest<Result>
 {
-
     public string CacheKey => LogsCacheKey.GetAllCacheKey;
     public CancellationTokenSource? SharedExpiryTokenSource => LogsCacheKey.SharedExpiryTokenSource();
-
 }
 
-public class ClearLogsCommandHandler :
-             IRequestHandler<ClearLogsCommand, Result>
+public class ClearLogsCommandHandler : IRequestHandler<ClearLogsCommand, Result>
 
 {
     private readonly IApplicationDbContext _context;
-    private readonly IMapper _mapper;
     private readonly IStringLocalizer<ClearLogsCommandHandler> _localizer;
     private readonly ILogger<ClearLogsCommandHandler> _logger;
+    private readonly IMapper _mapper;
 
     public ClearLogsCommandHandler(
         IApplicationDbContext context,
         IStringLocalizer<ClearLogsCommandHandler> localizer,
         ILogger<ClearLogsCommandHandler> logger,
-         IMapper mapper
-        )
+        IMapper mapper
+    )
     {
         _context = context;
         _localizer = localizer;
         _logger = logger;
         _mapper = mapper;
     }
+
     public async Task<Result> Handle(ClearLogsCommand request, CancellationToken cancellationToken)
     {
         _context.Loggers.RemoveRange(_context.Loggers);
         await _context.SaveChangesAsync(cancellationToken);
-        _logger.LogInformation("Logs was erased");
+        _logger.LogInformation("Logs have been erased");
         return await Result.SuccessAsync();
     }
-
 }
-
