+++ src/Blazor.Server.UI/Resources/Pages/Identity/Users/Profile.resx
@@ -124,7 +124,7 @@
     <value>Changed password successfully.</value>
   </data>
   <data name="Click upload a photo." xml:space="preserve">
-    <value>Click upload a photo.</value>
+      <value>Click to upload a photo.</value>
   </data>
   <data name="Confirm New Password" xml:space="preserve">
     <value>Confirm New Password</value>
@@ -133,7 +133,7 @@
     <value>Current Password</value>
   </data>
   <data name="current password is required!" xml:space="preserve">
-    <value>current password is required!</value>
+      <value>Current password is required!</value>
   </data>
   <data name="Display Name" xml:space="preserve">
     <value>Display Name</value>
@@ -145,7 +145,7 @@
     <value>New Password</value>
   </data>
   <data name="password is required!" xml:space="preserve">
-    <value>password is required!</value>
+      <value>Password is required!</value>
   </data>
   <data name="Phone Number" xml:space="preserve">
     <value>Phone Number</value>
@@ -160,7 +160,7 @@
     <value>Update successfully</value>
   </data>
   <data name="User Name" xml:space="preserve">
-    <value>User Name</value>
+      <value>Username</value>
   </data>
   <data name="Tenant Name" xml:space="preserve">
     <value>Tenant Name</value>
